renames extrators to extract to follow verb naming
