+++ dlt/pipeline/pipeline.py
@@ -19,7 +19,7 @@ from dlt.common.typing import DictStrAny, StrAny
 from dlt.common.utils import uniq_id, is_interactive
 from dlt.common.sources import DLT_METADATA_FIELD, TItem, with_table_name
 
-from dlt.extractors.extractor_storage import ExtractorStorageBase
+from dlt.extract.extractor_storage import ExtractorStorageBase
 from dlt.load.client_base import SqlClientBase, SqlJobClientBase
 from dlt.normalize.configuration import configuration as normalize_configuration
 from dlt.load.configuration import configuration as loader_configuration
